@@ -23,81 +23,81 @@
   },
   "license": "Apache-2.0",
   "dependencies": {
-    "@babel/polyfill": "~7.4.0",
-    "@babel/runtime": "~7.5.0",
+    "@babel/polyfill": "~7.6.0",
+    "@babel/runtime": "~7.6.0",
     "@fortawesome/fontawesome-svg-core": "~1.2.0",
     "@fortawesome/pro-light-svg-icons": "~5.7.0",
     "@fortawesome/pro-regular-svg-icons": "~5.7.0",
     "@fortawesome/pro-solid-svg-icons": "~5.7.0",
     "@fortawesome/react-fontawesome": "~0.1.0",
-    "@redux-saga/core": "~1.0.0",
+    "@redux-saga/core": "~1.1.0",
     "axios": "~0.19.0",
     "connected-react-router": "~6.5.0",
-    "history": "~4.9.0",
+    "history": "~4.10.0",
     "immutable": "4.0.0-rc.9",
-    "lattice": "0.62.1",
+    "lattice": "0.66.2",
     "lattice-auth": "0.15.0",
-    "lattice-sagas": "0.28.0",
-    "lattice-ui-kit": "0.9.4",
+    "lattice-sagas": "0.36.0",
+    "lattice-ui-kit": "0.16.0",
     "lodash": "~4.17.10",
     "loglevel": "~1.6.0",
-    "luxon": "~1.17.0",
+    "luxon": "~1.19.0",
     "polished": "~3.4.0",
-    "qs": "~6.7.0",
-    "react": "~16.9.0",
-    "react-dom": "~16.9.0",
+    "qs": "~6.9.0",
+    "react": "~16.10.0",
+    "react-dom": "~16.10.0",
     "react-redux": "~7.1.0",
-    "react-router": "~5.0.0",
-    "react-router-dom": "~5.0.0",
+    "react-router": "~5.1.0",
+    "react-router-dom": "~5.1.0",
     "react-select": "~3.0.0",
-    "react-sortable-hoc": "~1.9.0",
+    "react-sortable-hoc": "~1.10.0",
     "react-virtualized": "~9.21.0",
     "redux": "~4.0.0",
     "redux-immutable": "4.0.0",
     "redux-reqseq": "0.5.0",
-    "styled-components": "~4.3.0",
+    "styled-components": "~4.4.0",
     "validator": "~11.1.0"
   },
   "devDependencies": {
-    "@babel/cli": "~7.5.0",
-    "@babel/core": "~7.5.0",
+    "@babel/cli": "~7.6.0",
+    "@babel/core": "~7.6.0",
     "@babel/plugin-proposal-class-properties": "~7.5.0",
     "@babel/plugin-proposal-object-rest-spread": "~7.5.0",
-    "@babel/plugin-transform-runtime": "~7.5.0",
-    "@babel/preset-env": "~7.5.0",
+    "@babel/plugin-transform-runtime": "~7.6.0",
+    "@babel/preset-env": "~7.6.0",
     "@babel/preset-flow": "~7.0.0",
     "@babel/preset-react": "~7.0.0",
-    "@babel/register": "~7.4.0",
+    "@babel/register": "~7.6.0",
     "babel-eslint": "~10.0.0",
     "babel-jest": "~24.8.0",
     "babel-loader": "~8.0.0",
     "babel-plugin-styled-components": "~1.10.0",
-    "date-fns": "^1.30.1",
+    "date-fns": "^2.0.0",
     "enzyme": "~3.10.0",
     "enzyme-adapter-react-16": "~1.14.0",
     "enzyme-to-json": "~3.4.0",
-    "eslint": "~6.1.0",
+    "eslint": "~6.5.0",
     "eslint-config-airbnb": "~18.0.0",
     "eslint-import-resolver-webpack": "~0.11.0",
-    "eslint-loader": "~2.2.0",
-    "eslint-plugin-flowtype": "~4.2.0",
+    "eslint-loader": "~3.0.0",
+    "eslint-plugin-flowtype": "~4.3.0",
     "eslint-plugin-import": "~2.18.0",
-    "eslint-plugin-jest": "~22.15.0",
+    "eslint-plugin-jest": "~22.17.0",
     "eslint-plugin-jsx-a11y": "~6.2.0",
-    "eslint-plugin-react": "~7.14.0",
+    "eslint-plugin-react": "~7.15.0",
     "file-loader": "~4.2.0",
-    "flow-bin": "~0.105.0",
+    "flow-bin": "~0.109.0",
     "html-webpack-plugin": "~3.2.0",
     "jest": "~24.8.0",
     "jest-styled-components": "~6.3.0",
     "npm-run-all": "~4.1.0",
     "rimraf": "~3.0.0",
-    "stylelint": "~10.1.0",
-    "stylelint-config-standard": "~18.3.0",
+    "stylelint": "~11.0.0",
+    "stylelint-config-standard": "~19.0.0",
     "stylelint-config-styled-components": "~0.1.1",
     "stylelint-processor-styled-components": "~1.8.0",
     "uuid": "~3.3.0",
-    "webpack": "~4.39.0",
+    "webpack": "~4.41.0",
     "webpack-cli": "~3.3.0",
     "webpack-dev-server": "~3.8.0",
     "webpack-hot-middleware": "~2.25.0"
